Update zk.md
Hi, just noticed two typos in this section:

1. At second paragraph
'...which introduces a very efficienct zero-knowledge variant of KZG commitments...' should be '...which introduces a very efficient zero-knowledge variant of KZG commitments...',

2. At final paragraph
'...based on old work of Cramar and Damgard)',
Correction: This is likely a typo for Cramer. The well known cryptographer is Ivan Damgård, and his frequent collaborator on these topics is Ronald Cramer.

Thanks.